+++ examples/hover_shadow_html.rs
@@ -1,7 +1,7 @@
 fn main() {
     let mut s = poloto::plot("Demo: Hovering and shadows", "x", "y");
 
-    let x = (0..50).map(|x| (x as f32 / 50.0) * 10.0);
+    let x = (0..50).map(|x| (x as f64 / 50.0) * 10.0);
 
     s.line("cos", x.clone().map(|x| [x, x.cos()]));
     s.histogram("sin-3", x.clone().step_by(3).map(|x| [x, x.sin() - 3.]));
