+++ examples/unit.rs
@@ -5,9 +5,9 @@ struct Res{
     use_scientific:bool,
     precision:usize
 }
-fn write_data(low:f32,heigh:f32,step:f32)->Res{
+fn write_data(low:f64,heigh:f64,step:f64)->Res{
     //const SCIENCE: usize = 4;
-    //if a != 0.0 && a.abs().log10().floor().abs() > SCIENCE as f32 {
+    //if a != 0.0 && a.abs().log10().floor().abs() > SCIENCE as f64 {
     //    write!(fm, "{0:.1$e}", a, 2)?
     //} else {
     let k = (step.log10()).ceil();
