+++ examples/write_to_file.rs
@@ -5,7 +5,7 @@ fn main() {
         "This is the y label",
     );
 
-    let x = (0..50).map(|x| (x as f32 / 50.0) * 10.0);
+    let x = (0..50).map(|x| (x as f64 / 50.0) * 10.0);
 
     s.line("cos", x.clone().map(|x| [x, x.cos()]));
